Improvements for Spring Security module
- Added CORS support
- Fixed a bug with Username checking
- Moved the folders
- MAde sample a separate project. No more pom.xml inheritance
- Added CORS to the example as well
- Added securedRoute configuration
-
renamed:    spring-auth0-example/src/main/webapp/WEB-INF/web.xml -> examples/api-example/src/main/webapp/WEB-INF/web.xml
