+++ README.md
@@ -1,15 +1,51 @@
-spring-security-auth0
-=====================
+# Spring Security Auth0
 
-[![Build Status](https://travis-ci.org/calipho-sib/spring-security-auth0.svg?branch=master)](https://travis-ci.org/calipho-sib/spring-security-auth0)
+[Auth0](https://www.auth0.com) is a cloud service that provides a turn-key solution for authentication, authorization and Single Sign On. 
 
-An integration of spring security with [Auth0](http://auth0.com).
-This module is intended to be include in your existing spring mvc project.
+You can use  [Auth0](https://www.auth0.com) to add username/password authentication, support for enterprise identity like Active Directory or SAML and also for social identities like Google, Facebook or Salesforce among others to your web, API and mobile native apps. 
 
-This module was tested with Spring Security 3.2.4.
+## Learn how to use it
 
-Either see the [spring-auth0-core](spring-auth0-core/) for the integration instruction or look at the [spring-auth0-example](spring-auth0-example/) for a demo.
+[Please read this tutorial](https://docs.auth0.com/server-apis/java-spring-security) to learn how to use this SDK. It's exremely simple
 
+---
 
-### Questions / feedback / comments? 
-[Contact us](http://www.nextprot.org/contact/us)
+## What is Auth0?
+
+Auth0 helps you to:
+
+* Add authentication with [multiple authentication sources](https://docs.auth0.com/identityproviders), either social like **Google, Facebook, Microsoft Account, LinkedIn, GitHub, Twitter, Box, Salesforce, amont others**, or enterprise identity systems like **Windows Azure AD, Google Apps, Active Directory, ADFS or any SAML Identity Provider**. 
+* Add authentication through more traditional **[username/password databases](https://docs.auth0.com/mysql-connection-tutorial)**.
+* Add support for **[linking different user accounts](https://docs.auth0.com/link-accounts)** with the same user.
+* Support for generating signed [Json Web Tokens](https://docs.auth0.com/jwt) to call your APIs and **flow the user identity** securely.
+* Analytics of how, when and where users are logging in.
+* Pull data from other sources and add it to the user profile, through [JavaScript rules](https://docs.auth0.com/rules).
+
+## Create a free account in Auth0
+
+1. Go to [Auth0](http://developers.auth0.com) and click Sign Up.
+2. Use Google, GitHub or Microsoft Account to login.
+
+## License
+
+The MIT License (MIT)
+
+Copyright (c) 2013 AUTH10 LLC
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
