+++ examples/api-example/src/main/java/com/auth0/spring/security/auth0/SecuredPingController.java
@@ -1,0 +1,16 @@
+package com.auth0.spring.security.auth0;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class SecuredPingController {
+
+
+	@RequestMapping(value = "/secured/ping")
+	@ResponseBody
+	public String securedPing() {
+		return "All good. You only get this message if you're authenticated";
+	}
+}
