+++ examples/api-example/src/main/resources/auth0.properties
@@ -1,0 +1,4 @@
+auth0.clientSecret={CLIENT_SECRET}
+auth0.clientId={CLIENT_ID}
+auth0.domain={DOMAIN}
+auth0.securedRoute=/secured/**
