--- spring-auth0-core/src/main/java/sib/calipho/spring/security/auth0/Auth0TokenHelper.java
@@ -1,8 +1,0 @@
-package sib.calipho.spring.security.auth0;
-
-public interface Auth0TokenHelper<T> {
-
-	public String generateToken(T object, int expiration); 
-	public T decodeToken(String token);
-
-}
