--- spring-auth0-core/src/test/resources/application-context.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:security="http://www.springframework.org/schema/security"
-	xmlns:mvc="http://www.springframework.org/schema/mvc" xmlns:oxm="http://www.springframework.org/schema/oxm"
-	xmlns:lang="http://www.springframework.org/schema/lang"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans 
-		http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
- 		http://www.springframework.org/schema/context
-		http://www.springframework.org/schema/context/spring-context-3.2.xsd
-		http://www.springframework.org/schema/mvc 
-		http://www.springframework.org/schema/mvc/spring-mvc-3.2.xsd
-		http://www.springframework.org/schema/oxm 
-		http://www.springframework.org/schema/oxm/spring-oxm-3.2.xsd
-		http://www.springframework.org/schema/lang 
-		http://www.springframework.org/schema/lang/spring-lang-3.2.xsd
-		http://www.springframework.org/schema/security
-  	    http://www.springframework.org/schema/security/spring-security-3.2.xsd">
-
-	<import resource="classpath:auth0-security-context.xml" />
-
-	<bean id="auth0TokenHelper"
-		class="sib.calipho.spring.security.auth0.impl.Auth0TokenHelperImpl">
-		<property name="clientSecret" value="${auth0.clientSecret}"></property>
-		<property name="clientId" value="${auth0.clientId}"></property>
-	</bean>
-
-    <!-- Scan for spring annotated components -->
-    <context:component-scan base-package="sib.calipho"/>
-
-    <!-- Process annotations on registered beans like @Autowired... -->
-    <context:annotation-config/>
-
-    <!-- This tag registers the DefaultAnnotationHandlerMapping and
-         AnnotationMethodHandlerAdapter beans that are required for Spring MVC  -->
-    <mvc:annotation-driven/>
-
-	<!-- Resolves views selected for rendering by @Controllers to .jsp resources in the /WEB-INF/views directory -->
-	<bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-		<property name="prefix" value="/WEB-INF/views/" />
-		<property name="suffix" value=".jsp" />
-	</bean>
-	
-	<context:property-placeholder location="classpath:auth0.properties" />
-
-    <!-- This tag allows for mapping the DispatcherServlet to "/" -->
-    <mvc:default-servlet-handler/>
-
-    <!-- resources exclusions from servlet mapping -->
-    <mvc:resources mapping="/css/**" location="/css/"/>
-    <mvc:resources mapping="/js/**" location="/js/"/>
-
-
-</beans>
