--- spring-auth0-example/src/main/resources/log4j.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration debug="true"
-	xmlns:log4j='http://jakarta.apache.org/log4j/'>
-
-	<appender name="consoleAppender" class="org.apache.log4j.ConsoleAppender">
-		<layout class="org.apache.log4j.PatternLayout">
-			<param name="ConversionPattern" value="%d{dd MMM yyyy HH:mm:ss} %5p %c{1} - %m%n" />
-		</layout>
-	</appender>
-
-	<logger name="sib.calipho">
-		<level value="DEBUG" />
-	</logger>
-	
-	<logger name="org.springframework">
-		<level value="INFO" />
-	</logger>
-
-	<!-- root element must be below all others otherwise errors -->
-	<root>
-		<level value="INFO" />
-		<appender-ref ref="consoleAppender" />
-	</root>
-
-</log4j:configuration>
