--- spring-auth0-example/src/main/webapp/WEB-INF/web.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app version="2.5" xmlns="http://java.sun.com/xml/ns/javaee"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
-
-	<display-name>spring-auth0-example</display-name>
-
-	<context-param>
-		<param-name>contextConfigLocation</param-name>
-		<param-value>classpath:application-context.xml</param-value>
-	</context-param>
-
-
-    <!--  configure security --> 
-    <filter>
-            <filter-name>springSecurityFilterChain</filter-name>
-            <filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
-            <init-param>
-                    <param-name>contextAttribute</param-name>
-                    <param-value>org.springframework.web.servlet.FrameworkServlet.CONTEXT.auth0Api</param-value>
-            </init-param>
-    </filter>
-
-    <filter-mapping>
-            <filter-name>springSecurityFilterChain</filter-name>
-            <url-pattern>/*</url-pattern>
-    </filter-mapping>
-
- 	<listener>
-		<listener-class>org.springframework.web.context.ContextLoaderListener</listener-class>
-	</listener>
-	
-	
-	<!-- configure spring bootstrap for servlet -->
-	<servlet>
-		<servlet-name>auth0Api</servlet-name>
-		<servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-		<init-param>
-			<param-name>contextConfigLocation</param-name>
-			<param-value>/WEB-INF/spring/api-servlet.xml</param-value>
-		</init-param>
-		<load-on-startup>1</load-on-startup>
-	</servlet>
-
-	<servlet-mapping>
-		<servlet-name>auth0Api</servlet-name>
-		<url-pattern>/</url-pattern>
-	</servlet-mapping>
-
-</web-app>
