--- spring-auth0-example/src/main/webapp/css/app.css
@@ -1,82 +1,0 @@
-body {
-  font-family: "proxima-nova", sans-serif;
-  text-align: center;
-  font-size: 300%;
-  font-weight: 100;
-}
-input[type=checkbox],
-input[type=radio] {
-  position: absolute;
-  opacity: 0;
-}
-input[type=checkbox] + label,
-input[type=radio] + label {
-  display: inline-block;
-}
-input[type=checkbox] + label:before,
-input[type=radio] + label:before {
-  content: "";
-  display: inline-block;
-  vertical-align: -0.2em;
-  width: 1em;
-  height: 1em;
-  border: 0.15em solid #0074d9;
-  border-radius: 0.2em;
-  margin-right: 0.3em;
-  background-color: white;
-}
-input[type=radio] + label:before {
-  border-radius: 50%;
-}
-input[type=radio]:checked + label:before,
-input[type=checkbox]:checked + label:before {
-  background-color: #0074d9;
-  box-shadow: inset 0 0 0 0.15em white;
-}
-input[type=radio]:focus + label:before,
-input[type=checkbox]:focus + label:before {
-  outline: 0;
-}
-.btn:hover {
-  background-color: #44C7F4;
-}
-.btn:focus {
-  outline: none !important;
-}
-h1,
-h2,
-h3 {
-  font-weight: 100;
-}
-.home-description {
-  font-weight: 100;
-  margin: 100px 0;
-}
-h2 {
-  margin-top: 30px;
-  margin-bottom: 40px;
-  font-size: 200%;
-}
-#token {
-	font-size: 12px;
-	width: 800px;
-	height: 100px;
-	margin: 10px;
-}
-label {
-  font-size: 100%;
-  font-weight: 300;
-}
-.btn-next {
-  margin-top: 30px;
-}
-.answer {
-  width: 70%;
-  margin: auto;
-  text-align: left;
-  padding-left: 10%;
-  margin-bottom: 20px;
-}
-.login-page .login-box {
-  padding: 100px 0;
-}
