+++ src/test/resources/auth0.properties
@@ -1,0 +1,4 @@
+auth0.clientSecret=WU9VUl9TRUNSRVRfSUQ=
+auth0.clientId=YOUR_CLIENT_ID
+auth0.domain=YOUR_DOMAIN
+auth0.securedRoute=/secured/**
